Fix some bug with Calculator

+++ src/main/java/net/danh/dcore/Calculator/Calculator.java
@@ -1,5 +1,6 @@
 package net.danh.dcore.Calculator;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
@@ -47,7 +48,7 @@ public class Calculator {
                 }
             }
             if (s.charAt(sqrt + 1) == '{') {
-                double number = Double.parseDouble(calculator(s.substring(open + 1, close), -1));
+                double number = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(open + 1, close), -1))).doubleValue();
                 result = Math.sqrt(number);
             }
             return result;
@@ -61,16 +62,16 @@ public class Calculator {
             }
             for (int i = 0; i < pow.size(); i++) {
                 if (i == 0) {
-                    double a = Double.parseDouble(calculator(s.substring(0, pow.get(i)), -1));
+                    double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(0, pow.get(i)), -1))).doubleValue();
                     double b;
-                    if (i == pow.size() - 1) b = Double.parseDouble(calculator(s.substring(pow.get(0) + 1), -1));
-                    else b = Double.parseDouble(calculator(s.substring(pow.get(0) + 1, pow.get(1)), -1));
+                    if (i == pow.size() - 1) b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(pow.get(0) + 1), -1))).doubleValue();
+                    else b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(pow.get(0) + 1, pow.get(1)), -1))).doubleValue();
                     result += Math.pow(a, b);
                 } else if (i == pow.size() - 1) {
-                    double a = Double.parseDouble(calculator(s.substring(pow.get(i) + 1), -1));
+                    double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(pow.get(i) + 1), -1))).doubleValue();
                     result = Math.pow(result, a);
                 } else {
-                    double a = Double.parseDouble(calculator(s.substring(pow.get(i) + 1, pow.get(i + 1)), -1));
+                    double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(pow.get(i) + 1, pow.get(i + 1)), -1))).doubleValue();
                     result = Math.pow(result, a);
                 }
             }
@@ -85,23 +86,23 @@ public class Calculator {
             }
             for (int i = 0; i < percent.size(); i++) {
                 if (i == 0) {
-                    double a = Double.parseDouble(calculator(s.substring(0, percent.get(i)), -1));
+                    double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(0, percent.get(i)), -1))).doubleValue();
                     double b;
                     if (i == percent.size() - 1)
-                        b = Double.parseDouble(calculator(s.substring(percent.get(0) + 1), -1));
-                    else b = Double.parseDouble(calculator(s.substring(percent.get(0) + 1, percent.get(1)), -1));
+                        b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(percent.get(0) + 1), -1))).doubleValue();
+                    else b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(percent.get(0) + 1, percent.get(1)), -1))).doubleValue();
                     result += (b / 100) * a;
                 } else if (i == percent.size() - 1) {
-                    double a = Double.parseDouble(calculator(s.substring(percent.get(i) + 1), -1));
+                    double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(percent.get(i) + 1), -1))).doubleValue();
                     result = (a / 100) * result;
                 } else {
-                    double a = Double.parseDouble(calculator(s.substring(percent.get(i) + 1, percent.get(i + 1)), -1));
+                    double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(percent.get(i) + 1, percent.get(i + 1)), -1))).doubleValue();
                     result = (a / 100) * result;
                 }
             }
             return result;
         } else {
-            return Double.parseDouble(s);
+            return BigDecimal.valueOf(Long.parseLong(s)).doubleValue();
         }
     }
 
@@ -117,20 +118,20 @@ public class Calculator {
             }
             for (int i = 0; i < terms.size(); i++) {
                 if (i == 0) {
-                    double a = Double.parseDouble(calculator(s.substring(0, terms.get(i)), -1));
+                    double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(0, terms.get(i)), -1))).doubleValue();
                     double b;
-                    if (i == terms.size() - 1) b = Double.parseDouble(calculator(s.substring(terms.get(0) + 1), -1));
-                    else b = Double.parseDouble(calculator(s.substring(terms.get(0) + 1, terms.get(1)), -1));
+                    if (i == terms.size() - 1) b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(terms.get(0) + 1), -1))).doubleValue();
+                    else b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(terms.get(0) + 1, terms.get(1)), -1))).doubleValue();
                     if (s.charAt(terms.get(i)) == '*') result = a * b;
                     else if (s.charAt(terms.get(i)) == '/') result = a / b;
                     else if (s.charAt(terms.get(i)) == '%') result = a % b;
                 } else if (i == terms.size() - 1) {
-                    double b = Double.parseDouble(calculator(s.substring(terms.get(i) + 1), -1));
+                    double b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(terms.get(i) + 1), -1))).doubleValue();
                     if (s.charAt(terms.get(i)) == '*') result *= b;
                     else if (s.charAt(terms.get(i)) == '/') result /= b;
                     else if (s.charAt(terms.get(i)) == '%') result %= b;
                 } else {
-                    double b = Double.parseDouble(calculator(s.substring(terms.get(i) + 1, terms.get(i + 1)), -1));
+                    double b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(terms.get(i) + 1, terms.get(i + 1)), -1))).doubleValue();
                     if (s.charAt(terms.get(i)) == '*') result *= b;
                     else if (s.charAt(terms.get(i)) == '/') result /= b;
                     else if (s.charAt(terms.get(i)) == '%') result %= b;
@@ -138,7 +139,7 @@ public class Calculator {
             }
             return result;
         } else {
-            return Double.parseDouble(s);
+            return BigDecimal.valueOf(Long.parseLong(s)).doubleValue();
         }
     }
 
@@ -153,19 +154,19 @@ public class Calculator {
         }
         for (int i = 0; i < operators.size(); i++) {
             if (i == 0) {
-                double a = Double.parseDouble(calculator(s.substring(0, operators.get(i)), -1));
+                double a = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(0, operators.get(i)), -1))).doubleValue();
                 double b;
                 if (i == operators.size() - 1)
-                    b = Double.parseDouble(calculator(s.substring(operators.get(0) + 1), -1));
-                else b = Double.parseDouble(calculator(s.substring(operators.get(0) + 1, operators.get(1)), -1));
+                    b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(operators.get(0) + 1), -1))).doubleValue();
+                else b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(operators.get(0) + 1, operators.get(1)), -1))).doubleValue();
                 if (s.charAt(operators.get(i)) == '+') result = a + b;
                 else if (s.charAt(operators.get(i)) == '-') result = a - b;
             } else if (i == operators.size() - 1) {
-                double b = Double.parseDouble(calculator(s.substring(operators.get(i) + 1), -1));
+                double b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(operators.get(i) + 1), -1))).doubleValue();
                 if (s.charAt(operators.get(i)) == '+') result += b;
                 else if (s.charAt(operators.get(i)) == '-') result -= b;
             } else {
-                double b = Double.parseDouble(calculator(s.substring(operators.get(i) + 1, operators.get(i + 1)), -1));
+                double b = BigDecimal.valueOf(Long.parseLong(calculator(s.substring(operators.get(i) + 1, operators.get(i + 1)), -1))).doubleValue();
                 if (s.charAt(operators.get(i)) == '+') result += b;
                 else if (s.charAt(operators.get(i)) == '-') result -= b;
             }
